bonilla_banking/requirements.txt
asgiref==3.3.1
Django==3.1.7
django-address==0.2.5
django-mathfilters==1.0.0
python-git==2018.2.1
pytz==2021.1
Send2Trash==1.5.0
sqlparse==0.4.1